+++ src/naming-conventions/kebab-upper-case-naming-convention.ts
@@ -1,5 +1,5 @@
-import { Fn } from './../core/fn';
-import { NamingConvention } from './../core/naming-convention';
+import { Fn } from '../core/fn';
+import { NamingConvention } from '../core/naming-convention';
 
 /**
  * Kebab upper naming case.