+++ src/reference-handlers/index.ts
@@ -1,3 +1,3 @@
-export * from './direct.reference-handler';
-export * from './lead.reference-handler';
-export * from './path.reference-handler';
+export * from './direct-reference-handler';
+export * from './lead-reference-handler';
+export * from './path-reference-handler';